requirements.txt
pytest==7.3.2
PyYAML==6.0
allure-pytest=2.13.2
requests=2.31.0